fix(rpc): strict rippled parity for validator_info
- Error path: notValidator now returns rpcINVALID_PARAMS / "not a
  validator", matching rippled's not_validator_error() =
  make_param_error("not a validator"). SDK tooling pattern-matches
  on that exact wire shape.
- Drop the goXRPL-only `recent_validations` extension from the
  response. Tooling (xrpl.js, xrpl-py) expects rippled-shaped
  responses; a non-rippled field on the canonical handler invites
  parser drift across clients.
- Remove the now-unused archive plumbing exposed to RPC:
  ValidationArchiveLookup, ArchivedValidation,
  Services.ValidationArchive, and the
  ValidationArchiveAdapter wrapper. The on-disk archive itself is
  unaffected — it's still written by the consensus layer and
  available to future internal consumers.
- Relocate the handler test from internal/rpc to
  internal/rpc/handlers/validator_info_test.go using
  package handlers_test, matching the manifest_test.go convention
  for handler-scoped tests.
